feat: Implement mining via a loop
This now has enough logic to buy a ship if necessary, send it to an asteroid field, refuel it, mine and then stop when it's time to sell things (that's next).

I also added rate limiting (although I've never seen it hit it yet?) and a loop that knows how to wait until the next event.

Also fixed auth_token.txt to allow trailing newlines.