feat: Avoid mining with < 5 spots open
We might want to mine to completely full if we planned to transport to sell, but since we are currently
just selling to the local marketplace, there is no reason to mine when we're nearly full and would get
less than a full attempt's worth of units back.

I possibly should set it higher than 5.

Also fixed wrapping of one comment.