feat: Add (untested) arbitrage trader behavior
The arbitrage trader is probably wrong, but I'll figure that out when I try to run it next.

Also fixed a case where the cache a survey points to can become depleted and then mining will fail.  I no of no other way to see that it's depleted other than to try and catch the exception.

The data store seems to "leak" survey results (when I look at it it has more than one entry for a given waypointSymbol key), I'm not yet sure why or how to fix.