feat: Add WaypointCache, MarketCache and BehaviorManager
The purpose of this change is to allow better sharing of cached waypoint and market data between ships.  Previously waypoint objects were only (manually) cached for the starting system. Market objects were not cached at all.

This now makes all market and waypoint requests go through MarketCache and WaypointCache respectively, which live just as long as one loop through the ships and allows then ships spread across multiple systems to cache Waypoint and Market data.  This will become more important as we start to move between systems.

MarketCache.marketForSymbol is currently nullable, which may be a mistake.  I wanted to be able to cache that a waypoint doesn't have a marketplace, but i'm not sure that's the right way to do it.  All callers check that there is a marketplace at that waypoint before calling and we could possibly assert such (even if we still used null to cache non-marketplace-ness).  Caching a lack of marketplace is important for caching systems of markets I think?  That prevents us from doing a new cached lookup every time we ask for a whole system which has (commonly) non-market waypoints in it.

This also adds BehaviorManager which allows caching BehaviorState between loops.  This caching is not used yet, but the plumbing is (at least partially) there.  The first thing I intend to cache is Route data so ships can cache their intended route between loops and not have to re-plan (which can be expensive in terms of API calls).