feat: Make it possible to dynamically disable behaviors for one hour
This makes it possible for us to leave our arbitrage trader behavior on and have it wake up once an hour to check for deals. Once we make our arbitrage trader be able to look beyond a single system it should find more deals.

Our data store code is pretty janky and we should find a better way to handle typed data storage.