feat: move arbitrageTrader to multi-system navigation
Things don't fully seem to be working yet.  But I think this is worth checking in?
I renamed navigateTo to navigateToLocalWaypoint to be explicit that it does not handle navigations across systems.
I also added a navigation.dart file with beginRouteTo and made arbitrageTrader use it.

I fixed a bug in the trading logic where it was passing waypoints as "localWaypoint" that we did not have tradeGoods data for (no ship was there) and thus it was silently failing assuming no good were tradeable.

I also fixed another bug in the deal planning logic where it was asking for percentileForSellPrice rather than sellPriceForPercentile, I also renamed the methods to be more clear, and added some tests (finding yet another bug in crashing when asking for 100th percentile).

I also bumped the defaultMaxAge to 3 days from 1day and unified on a single constant to have one place to change.