Fix systemSymbolsInJumpRadius
As soon as I fixed systemSymbolsInJumpRadius the contract trader started doing things
which meant I had to fix the contract trader to not try to buy too many units for high-per-unit-cost contracts.

I need to add some tests, but I don't have time atm.

Also added the ability for WaypointCache.connectedSystems to check against the SystemCache.  I should invert the order so it prefers the system cache in a follow-up change.