fix: fix systemSymboslInJumpRadius to not duplicate symbols
There was a bug where further jump symbols could duplicate.  If the system was already queued for exploration in currentSystemsToJumpFrom and was seen again, it could get added to oneJumpFurther and then again added to currentSystemsToJumpFrom and yielded twice.  Fixed and tested.

I also refactored behavior manager to move it out of the loop and avoid calling getShips as many times.