chore: plumb AgentCache everywhere
This is in preparation for updating the agentCache when we get agent updating responses from the api.

I also made BuyShip alternate between ORE_HOUD and PROBE.

I pulled Agent fetching up into CLI in preparation for printing the current agent stats on boot.

Taught disableBehavior to take a timeout, and use that now in BuyShip