feat: include full route costs when planning trades
Wired in the new (very basic) route planner into the trading logic and verified that it now includes the cost of getting from the start to the source not just the source to destination in fuel and time costs.

Also fixed routing to/from the same place to return an empty route rather than null.