Split tradeVolume into expectedUnits and maxUnitsToBuy
This is to fix traders from never completing contracts because they
only have a tiny number of units left to complete and its not
ever worth it to haul a small number of units to complete them.

I also considered artificially raising the per-unit value for traders
but that would have caused them to buy goods which were overpriced.

This will just cause them to think they're going to fill up with 120, and
price for that, but then only by 10 but still complete the contract.