refactor: Use local BehaviorState
BehaviorState was originally introduced as side-loaded per-ship data
but we added passing it to behavior functions a while back.  This
change makes us use the local version of state in almost all cases.

Added BehaviorState.isComplete, but are not using it yet, it seemed
the most risky of the possible changes.