refactor: Use MultiJob for Miner
MinerJob is now stored on BehaviorState.  It's still computed
inside the job, but eventually we can move it to CentralCommand.

Also moved to using Cooldown on the Ship rather than
held in a side map.  This was a change with the latest server
update which I had not yet incorporated into my client.

I also had to then update several places where we were not
updating the Ship with the new Cooldown.  I also made sure
those called updateShip on ShipCache as well now.