feat: teach miners how to travel to sell
I had to teach findBestMarketToSell about minimum duration
as well as sell volume and units to sell.

The previous usage was wrong, since it didn't include unitsToSell
and would thus undervalue the extra trip time since it was
comparing against an absolute earnings per second, not one
per unit.

minimumDuration was necessary in order to get the miner to want to
travel to sell, since it accounts for the fact that the miner will
otherwise sit idle waiting for its reactor cooldown.

includeRoundTripCost was needed to model the return trip for miners.

This still does not account for fuel costs (which is wrong) and
could cause miners to end up traveling to sell at a loss when fuel
prices are high.