fix: hack around some db leak
We appear to cause the postgres process to consume multiple gb
over about 12-24hr periods.
The memory usage goes away as soon as we drop the connection.

I'm sure it's some bug in how I'm talking to postgres (possibly the
notification code?).  But for now just working around it with an
automated reconnect.