fix: simplify _waypointSymbolNeedingExploration
Made WaypointCache required, which fixed our latest null crasher
which happened when the chartingCache had values for a waypoint
but didn't have the traits cached necessary to fully reconstitute
the waypoint, violating a previous assumption of this code.

I removed the waypointCache = null codepath which simplifies things
and will also fix the crash.