fix: siphoning add FleetRole
Siphoning logic was never triggering because we
were chosing behavior based on registration.role
which is the same for both MINING_DRONE and
SIPHONING_DRONE.  So I added our own FleetRole
type which distinguishes.  It also was an opportunity
to clean up other places where we want to have a concept
of "what is this thing set up to do" which is separate from
"what is it doing now" or "what type did it have at purchase"
or "what frame does it have" or "what 'role' did the devs
assign it which may not actually be unique or correspond to
its current mounts or usage", etc.