feat: split MarketListingCache from MarketCache
This adds a MarketListingCache to hold the static "listing" data for
markets we've encountered.  Since the listing data is static, it can
be cached to disk and accessed synchronously, unlike MarketCache
which might hit the network.

I also split MarketCache and MarketListingCache out from
waypoint_cache.dart

This was all originally in service of trying to make
findWaypointPathWithinSystem synchronous so I could call it
from the RoutePlanner, but I decided not to finish that part of the
wiring yet in this change.

I also noticed that WaypointTraitCache was being loaded twice
and not shared, and so fixed that as part of adding
TradeGoodCache and ensuring that was shared between StaticCache
and the MarketListingCache.

I also changed the signature of all the navigation functions to
include Database and Caches, since they were already taking
half the caches as individual arguments and it was easier to just
assume they would need most of them, like how the behavior functions
and advanceShip work.

I also refactored planRoute into two helper functions (both of
which are currently stubs), one for planning jumps and a
second for planning travel within a waypoint.

I also added the concept of navDrift and refuel to the route planner
although I don't think either is implemented correctly yet.